feat(cq): updating cw as well
Ref: #1040
