+++ packages/api/src/external/carequality/command/cq-directory/get-cq-directory-entry-as-organization.ts
