+++ packages/api/src/external/carequality/index.ts
@@ -1,5 +1,5 @@
 import * as patient from "./patient";
-import * as organization from "./command/cq-directory/create-or-update-cq-organization";
+import * as organization from "./command/cq-organization/create-or-update-cq-organization";
 
 const cqCommands = {
   organization: {
