+++ packages/api/src/external/commonwell/command/cw-organization/parse-cw-organization.ts
@@ -1,0 +1,27 @@
+import { Organization } from "@metriport/commonwell-sdk";
+import { OID_PREFIX } from "@metriport/core/domain/oid";
+import { normalizeState, normalizeZipCodeNew } from "@metriport/shared";
+import { CwOrgDetails } from "../../shared";
+
+export function parseCWOrganization(org: Organization): CwOrgDetails {
+  const location = org.locations[0];
+  if (!location) throw new Error("Location not found");
+  return {
+    oid: org.organizationId.replace(OID_PREFIX, ""),
+    name: org.name,
+    data: {
+      name: org.name,
+      location: {
+        addressLine1: location.address1,
+        addressLine2: location.address2 ? location.address2 : undefined,
+        city: location.city,
+        state: normalizeState(location.state),
+        zip: normalizeZipCodeNew(location.postalCode),
+        country: location.country,
+      },
+      type: org.type,
+    },
+    active: org.isActive,
+    isObo: !!org.authorizationInformation,
+  };
+}
