+++ packages/api/src/external/hie/cross-hie-ids.ts
@@ -1,4 +1,4 @@
-import { getOrganizationIdsNotManagedBy } from "../carequality/command/cq-directory/get-cq-directory-organization";
+import { getOrganizationIdsNotManagedBy } from "../carequality/command/cq-directory/get-cq-directory-entry-as-organization";
 
 /**
  * Returns the list of Org OIDs that are not allowed to be used on CW.
